updating journal API's

+++ src/main/java/net/engineeringdigest/journalApp/service/JournalEntryService.java
@@ -41,12 +41,26 @@ public class JournalEntryService {
         return journalEntryRepo.findById(id);
     }
 
-    public void deleteId(ObjectId id, String userName){
-        User user = userService.findByUserName(userName);
-        user.getJournalEntries().removeIf(x->x.getId().equals(id));
-        userService.saveEntry(user);
-        journalEntryRepo.deleteById(id);
+    @Transactional
+    public boolean deleteId(ObjectId id, String userName){
+        boolean removed=false;
+        try {
+            User user = userService.findByUserName(userName);
+            removed = user.getJournalEntries().removeIf(x -> x.getId().equals(id));
+            if (removed){
+                userService.saveEntry(user);
+                journalEntryRepo.deleteById(id);
+            }
+        }catch (Exception e){
+            System.out.println(e);
+            throw new RuntimeException("An error occured while deleting the entry:"+e);
+        }
+        return removed;
     }
+
+//    public List<JournalEntry> findByUserName(String userName){
+//
+//    }
 
 
 }
